moved TransferExistingDepositsToHostZones
